161979: Add dockerfile to luke
