+++ docker-compose.yml
@@ -1,0 +1,29 @@
+version: '3'
+
+services:
+  web:
+    build: .
+    command: bash -c "python manage.py makemigrations && python manage.py collectstatic --noinput && python manage.py migrate && python manage.py runserver_plus 0.0.0.0:8000"
+    container_name: luke
+    environment:
+        DJANGO_SETTINGS_MODULE: 'luke.settings.local'
+    volumes:
+      - ./src:/src
+      - ./media:/media
+    ports:
+      - "8000:8000"
+    depends_on:
+      - db
+
+  db:
+    image: postgres:11
+    container_name: luke_db
+    volumes:
+      - postgres_data:/var/lib/postgresql/data/
+    environment:
+      - POSTGRES_USER=postgres
+      - POSTGRES_PASSWORD=postgres
+      - POSTGRES_DB=postgres
+
+volumes:
+  postgres_data:
