--- provision/templates/django/api/v1/routers.py
@@ -1,4 +1,0 @@
-# -*- coding: utf-8 -*-
-from ...core.api.routers import DefaultRouter
-
-router = DefaultRouter(trailing_slash=False)
