--- provision/templates/django/core/api/serializers/__init__.py
@@ -1,2 +1,0 @@
-# -*- coding: utf-8 -*-
-from .base import ModelSerializer  # noqa
