--- provision/templates/django/core/db/models/__init__.py
@@ -1,3 +1,0 @@
-# -*- coding: utf-8 -*-
-# flake8: noqa
-from .models import TimeStampedMixin, CatalogueMixin
