--- provision/templates/django/templates/index.html
@@ -1,10 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head>
-    <title>API</title>
-    <meta charset="UTF-8">
-</head>
-<body>
-    <p>Welcome!</p>
-</body>
-</html>
