--- provision/templates/django/urls.py
@@ -1,28 +1,0 @@
-"""${PROJECT_NAME} URL Configuration
- The `urlpatterns` list routes URLs to views. For more information please see:
-    https://docs.djangoproject.com/en/1.11/topics/http/urls/
-Examples:
-Function views
-    1. Add an import:  from my_app import views
-    2. Add a URL to urlpatterns:  url(r'^$', views.home, name='home')
-Class-based views
-    1. Add an import:  from other_app.views import Home
-    2. Add a URL to urlpatterns:  url(r'^$', Home.as_view(), name='home')
-Including another URLconf
-    1. Import the include() function: from django.conf.urls import url, include
-    2. Add a URL to urlpatterns:  url(r'^blog/', include('blog.urls'))
-"""
-from django.conf import settings
-from django.conf.urls import include, url
-from django.conf.urls.static import static
-from django.views.generic import TemplateView
-from '${PROJECT_NAME}'.api import urls as api_urls
-
-urlpatterns = [
-    url(
-        r'^api-auth/',
-        include('rest_framework.urls', namespace='rest_framework')
-    ),
-    url(r'^api/', include(api_urls, namespace='api')),
-    url(r'^$$', TemplateView.as_view(template_name="index.html")),
-] + static(settings.MEDIA_URL, document_root=settings.MEDIA_ROOT)
