--- provision/templates/nginx/local.conf
@@ -1,10 +1,0 @@
-server {
-    listen 80;
-
-    location / {
-        proxy_set_header Host $host;
-        proxy_set_header X-Real-IP $remote_addr;
-
-        proxy_pass http://127.0.0.1:8000;
-    }
-}
