--- provision/templates/yarn/package.json
@@ -1,6 +1,0 @@
-{
-  "name": "${PROJECT_NAME}",
-  "version": "0.0.1",
-  "dependencies": {},
-  "resolutions": {}
-}
