+++ src/package.json
@@ -1,0 +1,6 @@
+{
+    "name": "${PROJECT_NAME}",
+    "version": "0.0.1",
+    "dependencies": {},
+    "resolutions": {}
+}
