+++ src/requirements/base.txt
@@ -1,10 +1,10 @@
 # Core
-Django>=1.11,<2.0
+Django>=2.2,<2.3
 Pillow>=3.1,<3.2
 psycopg2-binary>=2.7,<2.8
 
 django-cors-headers>=1.1,<1.2
-django-extensions>=1.6,<1.7
+django-extensions>=2.1,<2.2
 django-filter==1.0.2
 djangorestframework==3.7.7
 djangorestframework-jsonapi==2.2.0
