+++ src/requirements/local.txt
@@ -1,4 +1,5 @@
 -r base.txt
 
 # Development tools
+django-debug-toolbar>=1.9,<1.10
 tox==2.6.0
