+++ src/tox.ini
@@ -1,0 +1,77 @@
+# Tox (http://tox.testrun.org/) is a tool for running tests
+# in multiple virtualenvs. This configuration file will run the
+# test suite on all supported python versions. To use it, "pip install tox"
+# and then run "tox" from this directory.
+[tox]
+skipsdist = true
+
+envlist =
+        {py36}-flake8
+        {py36}-django
+
+
+[testenv]
+deps =
+        -r{toxinidir}/requirements/testing.txt
+        pytest-cov
+        pytest-django
+        pytest-pythonpath
+        pytest-xdist
+
+setenv = DJANGO_SETTINGS_MODULE = luke.settings.testing
+
+commands =
+        py.test {posargs}
+        coverage erase
+
+
+[testenv:py36-flake8]
+deps =
+        flake8>=2.4,<2.6
+        flake8-import-order<0.8
+
+commands =
+        flake8 . {posargs}
+
+
+[flake8]
+
+exclude = migrations
+
+application-import-names = luke
+
+
+[pytest]
+django_find_project = false
+
+python_paths = .
+
+testpaths = .
+
+
+addopts =
+        --doctest-modules
+        -p no:warnings
+        --cov=luke .
+
+norecursedirs =
+        .*
+        requirements
+        luke/settings
+
+
+[coverage:run]
+source = .
+
+omit =
+        */migrations/*
+        */settings/*
+        */commands/*
+
+[coverage:report]
+
+fail_under = 30
+
+show_missing = True
+
+skip_covered = True
