+++ tox.ini
@@ -1,0 +1,87 @@
+# Tox (http://tox.testrun.org/) is a tool for running tests
+# in multiple virtualenvs. This configuration file will run the
+# test suite on all supported python versions. To use it, "pip install tox"
+# and then run "tox" from this directory.
+[tox]
+skipsdist = true
+
+envlist =
+        {py36}-flake8
+        {py36}-django
+
+
+
+
+[testenv]
+deps =
+        -r{toxinidir}/src/requirements/testing.txt
+        pytest-cov
+        pytest-django
+        pytest-pythonpath
+        pytest-xdist
+
+commands =
+        py.test {posargs}
+        coverage erase
+
+
+
+
+[testenv:py36-flake8]
+deps =
+        flake8>=2.4,<2.6
+        flake8-import-order<0.8
+
+commands =
+        flake8 src {posargs}
+
+
+
+
+[flake8]
+
+exclude = migrations
+
+application-import-names = luke
+
+
+
+
+[pytest]
+django_find_project = false
+
+python_paths = src/
+
+testpaths = src/
+
+DJANGO_SETTINGS_MODULE = luke.settings.testing
+
+addopts =
+        --doctest-modules
+        -p no:warnings
+        --cov=luke src/
+
+norecursedirs =
+        .*
+        src/requirements
+        src/luke/settings
+
+
+
+
+[coverage:run]
+source = src/
+
+omit =
+        */migrations/*
+        */settings/*
+
+
+
+
+[coverage:report]
+fail_under = 50
+
+show_missing = True
+
+skip_covered = True
